# src/api.yaml
openapi: 3.0.3
info:
  title: Ping API
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Server description
paths:
...
  /api/v1/github/{owner}/{repo}/issue/{issue_number}:
    parameters:
      - name: owner
        in: path
        required: true
        schema:
          type: string
      - name: repo
        in: path
        required: true
        schema:
          type: string
      - name: issue_number
        in: path
        required: true
        schema:
          type: number
    get:
      operationId: github
      responses:
        200:
          description: Returns id, title and body for a github issue
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                  id:
                    type: number
                  body:
                    type: string